fix/gateway: separate sse bean and tighten context reset
