+++ services/gateway/src/main/java/org/bumIntra/gateway/api/AdminResources.java
@@ -96,10 +96,6 @@ public class AdminResources {
             default -> "";
         };
 
-        if (prefix.isEmpty()) {
-            return subpath;
-        }
-
         if (subpath == null || subpath.isBlank()) {
             return prefix + "/";
         }
