+++ services/gateway/src/main/java/org/bumIntra/gateway/stream/SseProxyService.java
@@ -1,0 +1,90 @@
+package org.bumIntra.gateway.stream;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import org.bumIntra.gateway.obs.GatewayObserverDispatcher;
+import org.bumIntra.gateway.obs.event.GatewayRequestEnd;
+import org.bumIntra.gateway.security.GatewayRequestContext;
+import org.jboss.logging.Logger;
+import org.jboss.logging.MDC;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.StreamingOutput;
+
+@ApplicationScoped
+public class SseProxyService {
+
+    private static final Logger LOG = Logger.getLogger(SseProxyService.class);
+
+    @Inject
+    GatewayRequestContext grc;
+
+    @Inject
+    GatewayObserverDispatcher obs;
+
+    public Response buildStream(Response upstream) {
+        int status = upstream.getStatus();
+
+        if (status == 401 || status == 403 || status == 404) {
+            upstream.close();
+            emitRequestEnd(status, false);
+            return Response.status(status).build();
+        }
+
+        if (upstream.getStatusInfo().getFamily() != Response.Status.Family.SUCCESSFUL) {
+            upstream.close();
+            emitRequestEnd(Response.Status.BAD_GATEWAY.getStatusCode(), false);
+            return Response.status(Response.Status.BAD_GATEWAY)
+                    .entity("Failed to connect to upstream service")
+                    .build();
+        }
+
+        AtomicBoolean streamError = new AtomicBoolean(false);
+
+        StreamingOutput stream = output -> {
+            try (upstream; InputStream is = upstream.readEntity(InputStream.class)) {
+                byte[] buffer = new byte[8192];
+                int bytesRead;
+
+                while ((bytesRead = is.read(buffer)) != -1) {
+                    output.write(buffer, 0, bytesRead);
+                    output.flush();
+                }
+            } catch (IOException e) {
+                LOG.debugf("SSE stream closed or interrupted: requestId=%s, message=%s",
+                        grc.getRequestId(), e.getMessage());
+            } catch (Exception e) {
+                streamError.set(true);
+                LOG.errorf(e, "Unexpected SSE streaming failure: requestId=%s", grc.getRequestId());
+            } finally {
+                emitRequestEnd(status, !streamError.get());
+
+                MDC.clear();
+            }
+        };
+
+        return Response.ok(stream)
+                .type(MediaType.SERVER_SENT_EVENTS)
+                .header("Cache-Control", "no-cache")
+                .build();
+    }
+
+    private void emitRequestEnd(int status, boolean success) {
+        obs.onRequestEnd(new GatewayRequestEnd(
+                grc.getRequestId(),
+                status,
+                Duration.between(grc.getStartTime(), Instant.now()),
+                success,
+                Optional.ofNullable(grc.getErrorCode()),
+                Optional.ofNullable(grc.getServiceName()),
+                Optional.ofNullable(grc.getPathType())));
+    }
+}
